[EJIT] Harden AOT/runtime config-sync with compile-time and configure-time checks
Follow-up to the FUNC_SLOTS=64-vs-4096 production bug: audit both sides
(AOT passes / clang Sema and the runtime library) for every remaining
config value that exists in two or more independently-editable copies
with no hard check, and lock each one so a drift fails the build or the
configure instead of silently desynchronizing.

Compiler side (AOT passes + clang Sema):
- SECT_EJIT_BITCODE had the WRONG value ".ejit.bitcode" (dead but a
  landmine); fixed to ".ejit_bitcode", added SECT_EJIT_PERIOD, and the
  passes now emit sections via these constants instead of string
  literals. FN_REGISTER_SYMBOL centralizes the last hardcoded runtime
  entry name.
- Wrapper dim-cap error and Sema checks use EJIT_ICACHE_MAX_DIMS /
  MAX_PERIOD_ARR_IND_PARAMS / MAX_PERIOD_ARR_SIZE instead of literals;
  new static_asserts lock MAX_DIMS == MAX_PERIOD_ARR_IND_PARAMS and the
  DIM_SIZE power-of-2 stride in the AOT pass TU.

Runtime side:
- EJitSharedTaskPool.h FUNC_SLOTS fallback default 64u -> 4096u, so even
  non-CMake builds are safe by default (static_assert still backstops).
- EJitRuntime.cpp: new static_assert block locks the independently
  defined copies: ejit_log_level_t vs EJIT_LOG_LVL_* thresholds, the
  0xFE wrapper-timing sentinel vs every ejit_status_t value, the
  40-byte/8-align registry entry ABI, ejit_dim_pair_t vs EJitDimPair
  layout, and (shared-taskpool builds) kEJitMax* vs kEJitShared*
  capacities.
- EJitSharedTaskPool.cpp: D pow2 + MAX_DIMS == kEJitSharedMaxDims.
- EJitSrePlatform.cpp: 2MiB split granule and 4KiB seal page locked to
  kEJitSharedSplitGranule/kEJitSharedSealPage.
- ejit_activate/ejit_deactivate ABI widened uint8_t -> uint32_t and the
  runtime now REJECTS cellIdx >= kEJitMaxInstances (the old signature
  truncated at the call boundary and activated the wrong instance).
  gtest updated to cover the rejection.
- EJit.cpp: empty registry-section ranges get a VERBOSE diagnostic (the
  old behavior was a silent empty walk on hosted builds).

Build system:
- Configure-time numeric validation for all EJIT cache vars: pool size
  (2MiB multiple, >= 1), ptNo [1,255], buckets [1,254], queue capacity
  (pow2 >= 2), max funcIndex >= 1, worker stack/throttle >= 0, cache
  slots >= 1; a helper rejects non-numeric strings (CMake LESS/GREATER
  silently bypasses bounds on them).
- Unit-test/bench targets now forward the real cache-var values instead
  of hardcoded literals, so tuning a var keeps being tested.
- Stale -mllvm flag names in comments corrected to
  -print-ejit-icache-dim-size.

Verified: LLVMEJIT + LLVMEmbeddedJIT + clang build clean (diag ON,
shared taskpool, NO_RECLAIM); 4 InlineCache gtests + 82 EJITTaskPool
gtests pass; negative configure tests trigger the new FATAL_ERRORs
(non-2MiB pool, non-pow2 queue, out-of-range buckets, non-numeric
value).

+++ llvm/include/llvm/ExecutionEngine/EJIT/EJit.h
@@ -39,9 +39,12 @@ public:
 
   // Lifecycle — period-level (fans out to all arrays under periodName).
   // Returns false only in a taskpool build when periodName is not a registered
-  // lifecycle (no state is changed); always true in the legacy build.
-  bool activate(const std::string &periodName, uint8_t cellIdx);
-  bool deactivate(const std::string &periodName, uint8_t cellIdx);
+  // lifecycle (no state is changed), or when cellIdx >= kEJitMaxInstances;
+  // always true in the legacy build. cellIdx is uint32_t so an out-of-range
+  // instance index is rejected here instead of being truncated at the ABI
+  // boundary (see ejit_activate in EJitRuntime.h).
+  bool activate(const std::string &periodName, uint32_t cellIdx);
+  bool deactivate(const std::string &periodName, uint32_t cellIdx);
 
   bool activateAll(const std::string &periodName);
   bool deactivateAll(const std::string &periodName);
